Refactor task code into modules, use Redis for session-scoped mapping
Extracted ~730 lines of task-related code from the 3900+ line server.py into
focused modules under src/fastmcp/server/tasks/ and replaced the in-memory
task mapping with Redis-backed storage using Docket's Redis instance.

Changes:
- New handlers.py: Task execution handlers (handle_*_as_task functions)
- New protocol.py: Protocol handlers (tasks/get, tasks/result, tasks/list, tasks/cancel, tasks/delete)
- New converters.py: Result conversion logic (convert_*_result functions)
- Updated __init__.py: Exports public API
- Converted instance methods to standalone functions taking server param
- Used Context wrapper: `async with Context(fastmcp=server) as ctx:` for clean session_id access
- Replaced `_task_id_mapping` dict with Redis storage at `fastmcp:task:{session_id}:{client_task_id}`
- TTL: Docket's execution_ttl + 15 minutes (TASK_MAPPING_TTL_BUFFER_SECONDS constant)
- Removed `_task_id_mapping`, `_lock`, `set_state()`, `resolve_task_id()` from _temporary_mcp_shims
- Removed `_cancelled_tasks` tracking (now uses native Docket ExecutionState.CANCELLED)
- Updated pydocket requirement to >=0.13.2 for CANCELLED state support
- Updated test fixtures and metadata tests to use client protocol methods

Result: server.py reduced from 3938 → 3209 lines (19% reduction), all 147 tests passing
