+++ src/fastmcp/server/server.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import base64
 import inspect
 import json
 import re
@@ -22,10 +21,8 @@ from contextlib import (
     AbstractAsyncContextManager,
     AsyncExitStack,
     asynccontextmanager,
-    suppress,
 )
 from dataclasses import dataclass
-from datetime import timedelta
 from functools import partial
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Generic, Literal, cast, overload
@@ -33,21 +30,15 @@ from typing import TYPE_CHECKING, Any, Generic, Literal, cast, overload
 import anyio
 import httpx
 import mcp.types
-import pydantic_core
 import uvicorn
-from docket.execution import ExecutionState
 from mcp.server.lowlevel.helper_types import ReadResourceContents
 from mcp.server.lowlevel.server import LifespanResultT, NotificationOptions
 from mcp.server.stdio import stdio_server
-from mcp.shared.exceptions import McpError
 from mcp.types import (
-    INTERNAL_ERROR,
-    INVALID_PARAMS,
     Annotations,
     AnyFunction,
     CallToolRequestParams,
     ContentBlock,
-    ErrorData,
     GetPromptResult,
     ToolAnnotations,
 )
@@ -63,6 +54,9 @@ from starlette.routing import BaseRoute, Route
 
 import fastmcp
 import fastmcp.server
+
+# TODO SEP-1686: Import triggers monkey-patch of ClientRequest union for task methods
+import fastmcp.server.tasks._temporary_mcp_shims
 from fastmcp.exceptions import DisabledError, NotFoundError
 from fastmcp.mcp_config import MCPConfig
 from fastmcp.prompts import Prompt
@@ -79,16 +73,12 @@ from fastmcp.server.http import (
 )
 from fastmcp.server.low_level import LowLevelServer
 from fastmcp.server.middleware import Middleware, MiddlewareContext
-from fastmcp.server.tasks._temporary_mcp_shims import (
-    _lock as mcp_lock,
+from fastmcp.server.tasks.handlers import (
+    handle_prompt_as_task,
+    handle_resource_as_task,
+    handle_tool_as_task,
 )
-
-# TODO SEP-1686: Import triggers monkey-patch of ClientRequest union for task methods
-from fastmcp.server.tasks._temporary_mcp_shims import (
-    resolve_task_id,
-    set_state,
-)
-from fastmcp.server.tasks.keys import build_task_key
+from fastmcp.server.tasks.protocol import setup_task_protocol_handlers
 from fastmcp.settings import Settings
 from fastmcp.tools.tool import FunctionTool, Tool, ToolResult
 from fastmcp.tools.tool_manager import ToolManager
@@ -97,16 +87,6 @@ from fastmcp.utilities.cli import log_server_banner
 from fastmcp.utilities.components import FastMCPComponent
 from fastmcp.utilities.logging import get_logger, temporary_log_level
 from fastmcp.utilities.types import NotSet, NotSetT
-
-# Map Docket execution states to MCP task status strings
-DOCKET_TO_MCP_STATE: dict[ExecutionState, str] = {
-    ExecutionState.SCHEDULED: "submitted",
-    ExecutionState.QUEUED: "submitted",
-    ExecutionState.RUNNING: "working",
-    ExecutionState.COMPLETED: "completed",
-    ExecutionState.FAILED: "failed",
-    ExecutionState.CANCELLED: "cancelled",
-}
 
 if TYPE_CHECKING:
     from fastmcp.client import Client
@@ -630,68 +610,8 @@ class FastMCP(Generic[LifespanResultT]):
         self._mcp_server.request_handlers[mcp.types.ReadResourceRequest] = handler
 
     def _setup_task_protocol_handlers(self) -> None:
-        """
-        Register custom SEP-1686 task protocol handlers.
-
-        TODO SEP-1686: Remove this method when MCP SDK officially supports task protocol.
-        These handlers are registered directly because the SDK doesn't yet support
-        the tasks/get, tasks/result, and tasks/delete methods.
-        """
-        from fastmcp.server.tasks._temporary_mcp_shims import (
-            TasksCancelRequest,
-            TasksDeleteRequest,
-            TasksGetRequest,
-            TasksListRequest,
-            TasksResultRequest,
-        )
-
-        # Wrapper to adapt our dict-based handlers to MCP ServerResult format
-        async def tasks_get_handler(
-            req: TasksGetRequest,
-        ) -> mcp.types.ServerResult:
-            # Convert params to dict for our handler
-            params_dict = req.params.model_dump(by_alias=True, exclude_none=True)
-            result = await self._tasks_get_mcp(params_dict)
-            return mcp.types.ServerResult(root=result)  # type: ignore[arg-type]
-
-        async def tasks_result_handler(
-            req: TasksResultRequest,
-        ) -> mcp.types.ServerResult:
-            # Convert params to dict for our handler
-            params_dict = req.params.model_dump(by_alias=True, exclude_none=True)
-            result = await self._tasks_result_mcp(params_dict)
-            return mcp.types.ServerResult(root=result)  # type: ignore[arg-type]
-
-        async def tasks_delete_handler(
-            req: TasksDeleteRequest,
-        ) -> mcp.types.ServerResult:
-            # Convert params to dict for our handler
-            params_dict = req.params.model_dump(by_alias=True, exclude_none=True)
-            result = await self._tasks_delete_mcp(params_dict)
-            return mcp.types.ServerResult(root=result)  # type: ignore[arg-type]
-
-        async def tasks_list_handler(
-            req: TasksListRequest,
-        ) -> mcp.types.ServerResult:
-            # Convert params to dict for our handler
-            params_dict = req.params.model_dump(by_alias=True, exclude_none=True)
-            result = await self._tasks_list_mcp(params_dict)
-            return mcp.types.ServerResult(root=result)  # type: ignore[arg-type]
-
-        async def tasks_cancel_handler(
-            req: TasksCancelRequest,
-        ) -> mcp.types.ServerResult:
-            # Convert params to dict for our handler
-            params_dict = req.params.model_dump(by_alias=True, exclude_none=True)
-            result = await self._tasks_cancel_mcp(params_dict)
-            return mcp.types.ServerResult(root=result)  # type: ignore[arg-type]
-
-        # Register handlers directly
-        self._mcp_server.request_handlers[TasksGetRequest] = tasks_get_handler
-        self._mcp_server.request_handlers[TasksResultRequest] = tasks_result_handler
-        self._mcp_server.request_handlers[TasksDeleteRequest] = tasks_delete_handler
-        self._mcp_server.request_handlers[TasksListRequest] = tasks_list_handler
-        self._mcp_server.request_handlers[TasksCancelRequest] = tasks_cancel_handler
+        """Register custom SEP-1686 task protocol handlers."""
+        setup_task_protocol_handlers(self)
 
     async def _apply_middleware(
         self,
@@ -1323,8 +1243,8 @@ class FastMCP(Generic[LifespanResultT]):
                     tool = self._tool_manager._tools.get(key)
                     if tool and tool.task:
                         # Route to background execution
-                        return await self._handle_tool_as_task(
-                            key, arguments, task_meta
+                        return await handle_tool_as_task(
+                            self, key, arguments, task_meta
                         )
                     else:
                         # Graceful degradation per SEP-1686 spec:
@@ -1406,688 +1326,6 @@ class FastMCP(Generic[LifespanResultT]):
 
         raise NotFoundError(f"Unknown tool: {tool_name!r}")
 
-    async def _handle_tool_as_task(
-        self,
-        tool_name: str,
-        arguments: dict[str, Any],
-        task_meta: dict[str, Any],
-    ) -> mcp.types.CallToolResult:
-        """Handle tool execution as background task (SEP-1686).
-
-        Queues the user's actual function to Docket (preserving signature for DI),
-        stores raw return values, converts to MCP types on retrieval.
-
-        Args:
-            tool_name: Name of the tool to execute
-            arguments: Tool arguments
-            task_meta: Task metadata from request (contains taskId, keepAlive)
-
-        Returns:
-            CallToolResult: Task stub with task metadata in _meta
-        """
-        from fastmcp.server.dependencies import _current_docket, get_context
-
-        # Extract task parameters
-        client_task_id = task_meta["taskId"]
-        keep_alive = task_meta.get("keepAlive", 60000)
-
-        # Get session ID and Docket
-        ctx = get_context()
-        session_id = ctx.session_id
-
-        docket = _current_docket.get()
-        if docket is None:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message="Background tasks require Docket. Set FASTMCP_EXPERIMENTAL_ENABLE_DOCKET=true",
-                )
-            )
-
-        # Build full task key with embedded metadata
-        task_key = build_task_key(session_id, client_task_id, "tool", tool_name)
-
-        # Get the tool to access user's function
-        tool = await self.get_tool(tool_name)
-
-        # Set initial state and register mapping
-        await set_state(
-            task_key, "submitted", keep_alive=keep_alive, client_task_id=client_task_id
-        )
-
-        # Send notifications/tasks/created per SEP-1686 (mandatory)
-        # Send BEFORE queuing to avoid race where task completes before notification
-        notification = mcp.types.JSONRPCNotification(
-            jsonrpc="2.0",
-            method="notifications/tasks/created",
-            params={},  # Empty params per spec
-            _meta={  # taskId in _meta per spec
-                "modelcontextprotocol.io/related-task": {
-                    "taskId": client_task_id,
-                }
-            },
-        )
-
-        ctx = get_context()
-        with suppress(Exception):
-            # Don't let notification failures break task creation
-            await ctx.session.send_notification(notification)  # type: ignore[arg-type]
-
-        # Queue function to Docket (result storage via execution_ttl)
-        await docket.add(
-            tool.fn,  # type: ignore[attr-defined]
-            key=task_key,
-        )(**arguments)
-
-        # Return task stub
-        return mcp.types.CallToolResult(
-            content=[],
-            _meta={
-                "modelcontextprotocol.io/task": {
-                    "taskId": client_task_id,
-                    "status": "submitted",
-                }
-            },
-        )
-
-    async def _handle_prompt_as_task(
-        self,
-        prompt_name: str,
-        arguments: dict[str, Any] | None,
-        task_meta: dict[str, Any],
-    ) -> mcp.types.GetPromptResult:
-        """Handle prompt execution as background task (SEP-1686).
-
-        Queues the user's actual function to Docket (preserving signature for DI).
-
-        Args:
-            prompt_name: Name of the prompt to execute
-            arguments: Prompt arguments
-            task_meta: Task metadata from request (contains taskId, keepAlive)
-
-        Returns:
-            GetPromptResult: Task stub with task metadata in _meta
-        """
-        from fastmcp.server.dependencies import _current_docket, get_context
-
-        # Extract task parameters
-        client_task_id = task_meta["taskId"]
-        keep_alive = task_meta.get("keepAlive", 60000)
-
-        # Get session ID and Docket
-        ctx = get_context()
-        session_id = ctx.session_id
-
-        docket = _current_docket.get()
-        if docket is None:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message="Background tasks require Docket. Set FASTMCP_EXPERIMENTAL_ENABLE_DOCKET=true",
-                )
-            )
-
-        # Build full task key with embedded metadata
-        task_key = build_task_key(session_id, client_task_id, "prompt", prompt_name)
-
-        # Get the prompt
-        prompt = await self.get_prompt(prompt_name)
-
-        # Set initial state and register mapping
-        await set_state(
-            task_key, "submitted", keep_alive=keep_alive, client_task_id=client_task_id
-        )
-
-        # Send notifications/tasks/created per SEP-1686 (mandatory)
-        # Send BEFORE queuing to avoid race where task completes before notification
-        notification = mcp.types.JSONRPCNotification(
-            jsonrpc="2.0",
-            method="notifications/tasks/created",
-            params={},
-            _meta={
-                "modelcontextprotocol.io/related-task": {
-                    "taskId": client_task_id,
-                }
-            },
-        )
-        with suppress(Exception):
-            await ctx.session.send_notification(notification)  # type: ignore[arg-type]
-
-        # Queue function to Docket (result storage via execution_ttl)
-        await docket.add(
-            prompt.fn,  # type: ignore[attr-defined]
-            key=task_key,
-        )(**(arguments or {}))
-
-        # Return task stub
-        return mcp.types.GetPromptResult(
-            description="",
-            messages=[],
-            _meta={
-                "modelcontextprotocol.io/task": {
-                    "taskId": client_task_id,
-                    "status": "submitted",
-                }
-            },
-        )
-
-    async def _tasks_get_mcp(self, params: dict[str, Any]) -> dict[str, Any]:
-        """Handle MCP 'tasks/get' request (SEP-1686).
-
-        Args:
-            params: Request params containing taskId
-
-        Returns:
-            dict: Task status response
-        """
-
-        client_task_id = params.get("taskId")
-        if not client_task_id:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS, message="Missing required parameter: taskId"
-                )
-            )
-
-        # Resolve to full task key via mapping (for FastMCPTransport)
-        task_key = await resolve_task_id(client_task_id)
-        if task_key is None:
-            # Task not found - return unknown state per SEP-1686
-            return {
-                "taskId": client_task_id,
-                "status": "unknown",
-                "pollFrequency": 1000,
-                "_meta": {
-                    "modelcontextprotocol.io/related-task": {
-                        "taskId": client_task_id,
-                    }
-                },
-            }
-
-        # Get execution from Docket (use instance attribute for cross-task access)
-        docket = self._docket
-        if docket is None:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message="Background tasks require Docket",
-                )
-            )
-
-        execution = await docket.get_execution(task_key)
-        if execution is None:
-            # Task exists in mapping but no execution - unknown
-            return {
-                "taskId": client_task_id,
-                "status": "unknown",
-                "pollFrequency": 1000,
-                "_meta": {
-                    "modelcontextprotocol.io/related-task": {
-                        "taskId": client_task_id,
-                    }
-                },
-            }
-
-        # Sync state from Redis
-        await execution.sync()
-
-        # Map Docket state to MCP state
-        mcp_state = DOCKET_TO_MCP_STATE.get(execution.state, "unknown")
-
-        # Build response (use default keepAlive since we don't track per-task values)
-        response = {
-            "taskId": client_task_id,
-            "status": mcp_state,
-            "keepAlive": 60000,  # Default value
-            "pollFrequency": 1000,
-            "_meta": {
-                "modelcontextprotocol.io/related-task": {
-                    "taskId": client_task_id,
-                }
-            },
-        }
-
-        # Add error info if failed
-        if execution.state == ExecutionState.FAILED:
-            try:
-                await execution.get_result(timeout=timedelta(seconds=0))
-            except Exception as error:
-                response["error"] = str(error)
-
-        return response
-
-    async def _tasks_result_mcp(self, params: dict[str, Any]) -> Any:
-        """Handle MCP 'tasks/result' request (SEP-1686).
-
-        Converts raw task return values to MCP types based on task type.
-
-        Args:
-            params: Request params containing taskId
-
-        Returns:
-            MCP result (CallToolResult, GetPromptResult, or ReadResourceResult)
-        """
-        from fastmcp.server.tasks.keys import parse_task_key
-
-        client_task_id = params.get("taskId")
-        if not client_task_id:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS, message="Missing required parameter: taskId"
-                )
-            )
-
-        # Resolve to full task key via mapping (for FastMCPTransport)
-        from fastmcp.server.tasks._temporary_mcp_shims import resolve_task_id
-
-        task_key = await resolve_task_id(client_task_id)
-        if task_key is None:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Invalid taskId: {client_task_id} not found",
-                )
-            )
-
-        # Get execution from Docket (use instance attribute for cross-task access)
-        docket = self._docket
-        if docket is None:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message="Background tasks require Docket",
-                )
-            )
-
-        execution = await docket.get_execution(task_key)
-        if execution is None:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Invalid taskId: {client_task_id} not found",
-                )
-            )
-
-        # Sync state from Redis
-        await execution.sync()
-
-        # Check if completed
-        if execution.state not in (ExecutionState.COMPLETED, ExecutionState.FAILED):
-            mcp_state = DOCKET_TO_MCP_STATE.get(execution.state, "unknown")
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Task not completed yet (current state: {mcp_state})",
-                )
-            )
-
-        # Get result from Docket
-        try:
-            raw_value = await execution.get_result(timeout=timedelta(seconds=0))
-        except Exception as error:
-            # Task failed - return error result
-            return mcp.types.CallToolResult(
-                content=[mcp.types.TextContent(type="text", text=str(error))],
-                isError=True,
-                _meta={
-                    "modelcontextprotocol.io/related-task": {
-                        "taskId": client_task_id,
-                    }
-                },
-            )
-
-        # Parse task key to get type and component info
-        key_parts = parse_task_key(task_key)
-        task_type = key_parts["task_type"]
-
-        # Convert based on task type (pass client_task_id for metadata)
-        if task_type == "tool":
-            return await self._convert_raw_tool_result(
-                raw_value, key_parts["component_identifier"], client_task_id
-            )
-        elif task_type == "prompt":
-            return await self._convert_raw_prompt_result(
-                raw_value, key_parts["component_identifier"], client_task_id
-            )
-        elif task_type == "resource":
-            return await self._convert_raw_resource_result(
-                raw_value, key_parts["component_identifier"], client_task_id
-            )
-        else:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message=f"Internal error: Unknown task type: {task_type}",
-                )
-            )
-
-    async def _convert_raw_tool_result(
-        self, raw_value: Any, tool_name: str, client_task_id: str
-    ) -> mcp.types.CallToolResult:
-        """Convert raw tool return value to MCP CallToolResult.
-
-        Replicates the serialization logic from tool.run() to properly handle
-        output_schema, structured content, etc.
-
-        Args:
-            raw_value: The raw return value from user's tool function
-            tool_name: Name of the tool (to get output_schema and serializer)
-            client_task_id: Client task ID for related-task metadata
-
-        Returns:
-            CallToolResult with properly formatted content and structured content
-        """
-        from fastmcp.tools.tool import _convert_to_content
-
-        # Get the tool to access its configuration
-        tool = await self.get_tool(tool_name)
-
-        # Build related-task metadata
-        related_task_meta = {
-            "modelcontextprotocol.io/related-task": {
-                "taskId": client_task_id,
-            }
-        }
-
-        # If raw value is already ToolResult, use it directly
-        if isinstance(raw_value, ToolResult):
-            mcp_result = raw_value.to_mcp_result()
-            if isinstance(mcp_result, mcp.types.CallToolResult):
-                # Add metadata
-                mcp_result._meta = related_task_meta
-                return mcp_result
-            elif isinstance(mcp_result, tuple):
-                content, structured_content = mcp_result
-                return mcp.types.CallToolResult(
-                    content=content,
-                    structuredContent=structured_content,
-                    _meta=related_task_meta,
-                )
-            else:
-                return mcp.types.CallToolResult(
-                    content=mcp_result, _meta=related_task_meta
-                )
-
-        # Convert raw value to content blocks
-        unstructured_result = _convert_to_content(raw_value, serializer=tool.serializer)
-
-        # Handle structured content creation (same logic as tool.run())
-        structured_content = None
-
-        if tool.output_schema is None:
-            # Try to serialize as dict for structured content
-            try:
-                sc = pydantic_core.to_jsonable_python(raw_value)
-                if isinstance(sc, dict):
-                    structured_content = sc
-            except pydantic_core.PydanticSerializationError:
-                pass
-        else:
-            # Has output_schema - convert to JSON-able types
-            jsonable_value = pydantic_core.to_jsonable_python(raw_value)
-            wrap_result = tool.output_schema.get("x-fastmcp-wrap-result")
-            structured_content = (
-                {"result": jsonable_value} if wrap_result else jsonable_value
-            )
-
-        return mcp.types.CallToolResult(
-            content=unstructured_result,
-            structuredContent=structured_content,
-            _meta=related_task_meta,
-        )
-
-    async def _convert_raw_prompt_result(
-        self, raw_value: Any, prompt_name: str, client_task_id: str
-    ) -> mcp.types.GetPromptResult:
-        """Convert raw prompt return value to MCP GetPromptResult.
-
-        The user function returns raw values (strings, dicts, lists) that need
-        to be converted to PromptMessage objects.
-
-        Args:
-            raw_value: The raw return value from user's prompt function
-            prompt_name: Name of the prompt
-            client_task_id: Client task ID for related-task metadata
-
-        Returns:
-            GetPromptResult with properly formatted messages
-        """
-        from fastmcp.prompts.prompt import PromptMessage
-
-        # Get the prompt for metadata
-        prompt = await self.get_prompt(prompt_name)
-
-        # Normalize to list
-        if not isinstance(raw_value, list | tuple):
-            raw_value = [raw_value]
-
-        # Convert to PromptMessages
-        messages: list[mcp.types.PromptMessage] = []
-        for msg in raw_value:
-            if isinstance(msg, PromptMessage):
-                messages.append(msg.to_mcp())
-            elif isinstance(msg, str):
-                messages.append(
-                    mcp.types.PromptMessage(
-                        role="user",
-                        content=mcp.types.TextContent(type="text", text=msg),
-                    )
-                )
-            elif isinstance(msg, dict):
-                messages.append(mcp.types.PromptMessage.model_validate(msg))
-            else:
-                raise ValueError(f"Invalid message type: {type(msg)}")
-
-        return mcp.types.GetPromptResult(
-            description=prompt.description or "",
-            messages=messages,
-            _meta={
-                "modelcontextprotocol.io/related-task": {
-                    "taskId": client_task_id,
-                }
-            },
-        )
-
-    async def _convert_raw_resource_result(
-        self, raw_value: Any, uri: str, client_task_id: str
-    ) -> dict[str, Any]:
-        """Convert raw resource return value to MCP resource contents dict.
-
-        Args:
-            raw_value: The raw return value from user's resource function (str or bytes)
-            uri: Resource URI (for the contents response)
-            client_task_id: Client task ID for related-task metadata
-
-        Returns:
-            Dict with 'contents' key containing list of resource contents
-        """
-        # Build related-task metadata
-        related_task_meta = {
-            "modelcontextprotocol.io/related-task": {
-                "taskId": client_task_id,
-            }
-        }
-
-        # Resources return str or bytes directly
-        if isinstance(raw_value, str):
-            return {
-                "contents": [
-                    {
-                        "uri": uri,
-                        "text": raw_value,
-                        "mimeType": "text/plain",
-                    }
-                ],
-                "_meta": related_task_meta,
-            }
-        elif isinstance(raw_value, bytes):
-            return {
-                "contents": [
-                    {
-                        "uri": uri,
-                        "blob": base64.b64encode(raw_value).decode(),
-                        "mimeType": "application/octet-stream",
-                    }
-                ],
-                "_meta": related_task_meta,
-            }
-        else:
-            # Fallback: convert to JSON string
-            return {
-                "contents": [
-                    {
-                        "uri": uri,
-                        "text": json.dumps(raw_value),
-                        "mimeType": "application/json",
-                    }
-                ],
-                "_meta": related_task_meta,
-            }
-
-    async def _tasks_list_mcp(self, params: dict[str, Any]) -> dict[str, Any]:
-        """Handle MCP 'tasks/list' request (SEP-1686).
-
-        Note: With client-side tracking, this returns minimal info.
-
-        Args:
-            params: Request params (cursor, limit)
-
-        Returns:
-            dict: Response with tasks list and pagination
-        """
-        # Return empty list - client tracks tasks locally
-        # Note: tasks/list is not task-specific, so _meta doesn't have taskId
-        return {"tasks": [], "nextCursor": None, "_meta": {}}
-
-    async def _tasks_cancel_mcp(self, params: dict[str, Any]) -> dict[str, Any]:
-        """Handle MCP 'tasks/cancel' request (SEP-1686).
-
-        Cancels a running task, transitioning it to cancelled state.
-
-        Args:
-            params: Request params containing taskId
-
-        Returns:
-            dict: Task status response showing cancelled state
-        """
-        client_task_id = params.get("taskId")
-        if not client_task_id:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS, message="Missing required parameter: taskId"
-                )
-            )
-
-        # Resolve to full task key via mapping
-        task_key = await resolve_task_id(client_task_id)
-        if task_key is None:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Invalid taskId: {client_task_id} not found",
-                )
-            )
-
-        docket = self._docket
-        if docket is None:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message="Background tasks require Docket",
-                )
-            )
-
-        # Check if task exists
-        execution = await docket.get_execution(task_key)
-        if execution is None:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Invalid taskId: {client_task_id} not found",
-                )
-            )
-
-        # Cancel via Docket (now sets CANCELLED state natively)
-        await docket.cancel(task_key)
-
-        # Return task status with cancelled state
-        return {
-            "taskId": client_task_id,
-            "status": "cancelled",
-            "pollFrequency": 1000,
-            "_meta": {
-                "modelcontextprotocol.io/related-task": {
-                    "taskId": client_task_id,
-                }
-            },
-        }
-
-    async def _tasks_delete_mcp(self, params: dict[str, Any]) -> dict[str, Any]:
-        """Handle MCP 'tasks/delete' request (SEP-1686).
-
-        Deletion is discretionary - we allow deletion of any task.
-
-        Args:
-            params: Request params containing taskId
-
-        Returns:
-            dict: Empty response with related-task metadata
-        """
-        client_task_id = params.get("taskId")
-        if not client_task_id:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS, message="Missing required parameter: taskId"
-                )
-            )
-
-        # Resolve task ID and delete
-        from fastmcp.server.tasks._temporary_mcp_shims import _task_id_mapping
-
-        task_key = await resolve_task_id(client_task_id)
-        if task_key is None:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Invalid taskId: {client_task_id} not found",
-                )
-            )
-
-        docket = self._docket
-        if docket is None:
-            raise McpError(
-                ErrorData(
-                    code=INTERNAL_ERROR,
-                    message="Background tasks require Docket",
-                )
-            )
-
-        # Check if task exists
-        execution = await docket.get_execution(task_key)
-        if execution is None:
-            raise McpError(
-                ErrorData(
-                    code=INVALID_PARAMS,
-                    message=f"Invalid taskId: {client_task_id} not found",
-                )
-            )
-
-        # Cancel via Docket (Docket handles cleanup via TTL)
-        await docket.cancel(task_key)
-
-        # Remove from task ID mapping
-        async with mcp_lock:
-            _task_id_mapping.pop(client_task_id, None)
-
-        # Return empty response with related-task metadata
-        return {
-            "_meta": {
-                "modelcontextprotocol.io/related-task": {
-                    "taskId": client_task_id,
-                }
-            }
-        }
-
     async def _read_resource_mcp(
         self, uri: AnyUrl | str
     ) -> list[ReadResourceContents] | mcp.types.ServerResult:
@@ -2118,75 +1356,8 @@ class FastMCP(Generic[LifespanResultT]):
                     try:
                         resource = await self._resource_manager.get_resource(uri)
                         if resource and resource.task:
-                            from fastmcp.server.dependencies import (
-                                _current_docket,
-                                get_context,
-                            )
-
-                            # Extract task parameters
-                            client_task_id = task_meta["taskId"]
-                            keep_alive = task_meta.get("keepAlive", 60000)
-
-                            # Get session ID and Docket
-                            ctx = get_context()
-                            session_id = ctx.session_id
-
-                            docket = _current_docket.get()
-                            if docket is None:
-                                raise McpError(
-                                    ErrorData(
-                                        code=INTERNAL_ERROR,
-                                        message="Background tasks require Docket",
-                                    )
-                                )
-
-                            # Build full task key with embedded metadata (use original URI)
-                            task_key = build_task_key(
-                                session_id, client_task_id, "resource", str(uri)
-                            )
-
-                            # Set initial state and register mapping
-                            await set_state(
-                                task_key,
-                                "submitted",
-                                keep_alive=keep_alive,
-                                client_task_id=client_task_id,
-                            )
-
-                            # Send notifications/tasks/created per SEP-1686 (mandatory)
-                            # Send BEFORE queuing to avoid race where task completes before notification
-                            from contextlib import suppress
-
-                            notification = mcp.types.JSONRPCNotification(
-                                jsonrpc="2.0",
-                                method="notifications/tasks/created",
-                                params={},
-                                _meta={
-                                    "modelcontextprotocol.io/related-task": {
-                                        "taskId": client_task_id,
-                                    }
-                                },
-                            )
-                            with suppress(Exception):
-                                await ctx.session.send_notification(notification)  # type: ignore[arg-type]
-
-                            # Queue function to Docket (result storage via execution_ttl)
-                            await docket.add(
-                                resource.fn,  # type: ignore[attr-defined]
-                                key=task_key,
-                            )()
-
-                            # Return task stub
-                            return mcp.types.ServerResult(
-                                mcp.types.ReadResourceResult(
-                                    contents=[],
-                                    _meta={
-                                        "modelcontextprotocol.io/task": {
-                                            "taskId": client_task_id,
-                                            "status": "submitted",
-                                        }
-                                    },
-                                )
+                            return await handle_resource_as_task(
+                                self, str(uri), resource, task_meta
                             )
                     except NotFoundError:
                         # Resource not found - will fail in normal path
@@ -2305,8 +1476,8 @@ class FastMCP(Generic[LifespanResultT]):
                 if task_meta and self.settings.experimental.enable_tasks:
                     prompt = self._prompt_manager._prompts.get(name)
                     if prompt and prompt.task:
-                        return await self._handle_prompt_as_task(
-                            name, arguments, task_meta
+                        return await handle_prompt_as_task(
+                            self, name, arguments, task_meta
                         )
 
                 return await self._get_prompt_middleware(name, arguments)
