+++ src/fastmcp/server/tasks/handlers.py
@@ -1,0 +1,272 @@
+"""SEP-1686 task execution handlers.
+
+Handles queuing tool/prompt/resource executions to Docket as background tasks.
+"""
+
+from __future__ import annotations
+
+from contextlib import suppress
+from typing import TYPE_CHECKING, Any
+
+import mcp.types
+from mcp.shared.exceptions import McpError
+from mcp.types import INTERNAL_ERROR, ErrorData
+
+from fastmcp.server.dependencies import _current_docket, get_context
+from fastmcp.server.tasks.keys import build_task_key
+
+if TYPE_CHECKING:
+    from fastmcp.server.server import FastMCP
+
+# Redis mapping TTL buffer: Add 15 minutes to Docket's execution_ttl
+TASK_MAPPING_TTL_BUFFER_SECONDS = 15 * 60
+
+
+async def handle_tool_as_task(
+    server: FastMCP,
+    tool_name: str,
+    arguments: dict[str, Any],
+    task_meta: dict[str, Any],
+) -> mcp.types.CallToolResult:
+    """Handle tool execution as background task (SEP-1686).
+
+    Queues the user's actual function to Docket (preserving signature for DI),
+    stores raw return values, converts to MCP types on retrieval.
+
+    Args:
+        server: FastMCP server instance
+        tool_name: Name of the tool to execute
+        arguments: Tool arguments
+        task_meta: Task metadata from request (contains taskId, keepAlive)
+
+    Returns:
+        CallToolResult: Task stub with task metadata in _meta
+    """
+    # Extract task parameters
+    client_task_id = task_meta["taskId"]
+
+    # Get session ID and Docket
+    ctx = get_context()
+    session_id = ctx.session_id
+
+    docket = _current_docket.get()
+    if docket is None:
+        raise McpError(
+            ErrorData(
+                code=INTERNAL_ERROR,
+                message="Background tasks require Docket. Set FASTMCP_EXPERIMENTAL_ENABLE_DOCKET=true",
+            )
+        )
+
+    # Build full task key with embedded metadata
+    task_key = build_task_key(session_id, client_task_id, "tool", tool_name)
+
+    # Get the tool to access user's function
+    tool = await server.get_tool(tool_name)
+
+    # Store task key mapping in Redis for protocol handlers to look up
+    redis_key = f"fastmcp:task:{session_id}:{client_task_id}"
+    ttl_seconds = int(
+        docket.execution_ttl.total_seconds() + TASK_MAPPING_TTL_BUFFER_SECONDS
+    )
+    async with docket.redis() as redis:
+        await redis.set(redis_key, task_key, ex=ttl_seconds)
+
+    # Send notifications/tasks/created per SEP-1686 (mandatory)
+    # Send BEFORE queuing to avoid race where task completes before notification
+    notification = mcp.types.JSONRPCNotification(
+        jsonrpc="2.0",
+        method="notifications/tasks/created",
+        params={},  # Empty params per spec
+        _meta={  # taskId in _meta per spec
+            "modelcontextprotocol.io/related-task": {
+                "taskId": client_task_id,
+            }
+        },
+    )
+
+    ctx = get_context()
+    with suppress(Exception):
+        # Don't let notification failures break task creation
+        await ctx.session.send_notification(notification)  # type: ignore[arg-type]
+
+    # Queue function to Docket (result storage via execution_ttl)
+    await docket.add(
+        tool.fn,  # type: ignore[attr-defined]
+        key=task_key,
+    )(**arguments)
+
+    # Return task stub
+    return mcp.types.CallToolResult(
+        content=[],
+        _meta={
+            "modelcontextprotocol.io/task": {
+                "taskId": client_task_id,
+                "status": "submitted",
+            }
+        },
+    )
+
+
+async def handle_prompt_as_task(
+    server: FastMCP,
+    prompt_name: str,
+    arguments: dict[str, Any] | None,
+    task_meta: dict[str, Any],
+) -> mcp.types.GetPromptResult:
+    """Handle prompt execution as background task (SEP-1686).
+
+    Queues the user's actual function to Docket (preserving signature for DI).
+
+    Args:
+        server: FastMCP server instance
+        prompt_name: Name of the prompt to execute
+        arguments: Prompt arguments
+        task_meta: Task metadata from request (contains taskId, keepAlive)
+
+    Returns:
+        GetPromptResult: Task stub with task metadata in _meta
+    """
+    # Extract task parameters
+    client_task_id = task_meta["taskId"]
+
+    # Get session ID and Docket
+    ctx = get_context()
+    session_id = ctx.session_id
+
+    docket = _current_docket.get()
+    if docket is None:
+        raise McpError(
+            ErrorData(
+                code=INTERNAL_ERROR,
+                message="Background tasks require Docket. Set FASTMCP_EXPERIMENTAL_ENABLE_DOCKET=true",
+            )
+        )
+
+    # Build full task key with embedded metadata
+    task_key = build_task_key(session_id, client_task_id, "prompt", prompt_name)
+
+    # Get the prompt
+    prompt = await server.get_prompt(prompt_name)
+
+    # Store task key mapping in Redis for protocol handlers to look up
+    redis_key = f"fastmcp:task:{session_id}:{client_task_id}"
+    ttl_seconds = int(
+        docket.execution_ttl.total_seconds() + TASK_MAPPING_TTL_BUFFER_SECONDS
+    )
+    async with docket.redis() as redis:
+        await redis.set(redis_key, task_key, ex=ttl_seconds)
+
+    # Send notifications/tasks/created per SEP-1686 (mandatory)
+    # Send BEFORE queuing to avoid race where task completes before notification
+    notification = mcp.types.JSONRPCNotification(
+        jsonrpc="2.0",
+        method="notifications/tasks/created",
+        params={},
+        _meta={
+            "modelcontextprotocol.io/related-task": {
+                "taskId": client_task_id,
+            }
+        },
+    )
+    with suppress(Exception):
+        await ctx.session.send_notification(notification)  # type: ignore[arg-type]
+
+    # Queue function to Docket (result storage via execution_ttl)
+    await docket.add(
+        prompt.fn,  # type: ignore[attr-defined]
+        key=task_key,
+    )(**(arguments or {}))
+
+    # Return task stub
+    return mcp.types.GetPromptResult(
+        description="",
+        messages=[],
+        _meta={
+            "modelcontextprotocol.io/task": {
+                "taskId": client_task_id,
+                "status": "submitted",
+            }
+        },
+    )
+
+
+async def handle_resource_as_task(
+    server: FastMCP,
+    uri: str,
+    resource,  # Resource or ResourceTemplate
+    task_meta: dict[str, Any],
+) -> mcp.types.ServerResult:
+    """Handle resource read as background task (SEP-1686).
+
+    Queues the user's actual function to Docket.
+
+    Args:
+        server: FastMCP server instance
+        uri: Resource URI
+        resource: Resource or ResourceTemplate object
+        task_meta: Task metadata from request (contains taskId, keepAlive)
+
+    Returns:
+        ServerResult with ReadResourceResult stub
+    """
+    # Extract task parameters
+    client_task_id = task_meta["taskId"]
+
+    # Get session ID and Docket
+    ctx = get_context()
+    session_id = ctx.session_id
+
+    docket = _current_docket.get()
+    if docket is None:
+        raise McpError(
+            ErrorData(
+                code=INTERNAL_ERROR,
+                message="Background tasks require Docket",
+            )
+        )
+
+    # Build full task key with embedded metadata (use original URI)
+    task_key = build_task_key(session_id, client_task_id, "resource", str(uri))
+
+    # Store task key mapping in Redis for protocol handlers to look up
+    redis_key = f"fastmcp:task:{session_id}:{client_task_id}"
+    ttl_seconds = int(
+        docket.execution_ttl.total_seconds() + TASK_MAPPING_TTL_BUFFER_SECONDS
+    )
+    async with docket.redis() as redis:
+        await redis.set(redis_key, task_key, ex=ttl_seconds)
+
+    # Send notifications/tasks/created per SEP-1686 (mandatory)
+    # Send BEFORE queuing to avoid race where task completes before notification
+    notification = mcp.types.JSONRPCNotification(
+        jsonrpc="2.0",
+        method="notifications/tasks/created",
+        params={},
+        _meta={
+            "modelcontextprotocol.io/related-task": {
+                "taskId": client_task_id,
+            }
+        },
+    )
+    with suppress(Exception):
+        await ctx.session.send_notification(notification)  # type: ignore[arg-type]
+
+    # Queue function to Docket (result storage via execution_ttl)
+    await docket.add(
+        resource.fn,  # type: ignore[attr-defined]
+        key=task_key,
+    )()
+
+    # Return task stub
+    return mcp.types.ServerResult(
+        mcp.types.ReadResourceResult(
+            contents=[],
+            _meta={
+                "modelcontextprotocol.io/task": {
+                    "taskId": client_task_id,
+                    "status": "submitted",
+                }
+            },
+        )
+    )
