+++ tests/client/tasks/conftest.py
@@ -13,8 +13,3 @@ async def enable_docket_and_tasks():
         experimental__enable_tasks=True,
     ):
         yield
-
-        # Clean up MCP shim storage after each test
-        from fastmcp.server.tasks import _temporary_mcp_shims
-
-        _temporary_mcp_shims._task_id_mapping.clear()
